requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-binance>=1.0.19
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
vaderSentiment>=3.3.2
praw>=7.7.1
feedparser>=6.0.11
ta>=0.10.2
python-dotenv>=1.0.0
streamlit>=1.31.0
